Add newbs-localsrc class for fetching source from the local workspace
Look in NEWBS_SRCDIR (NEWBSROOT/newbs-source) for NEWBS_SRCNAME and
fetch that.

A side-effect is that the .git directory's symlinks to ../../.repo are
copied too, and a warning about not being able to checksum the (broken)
link .git/shallow gets printed.
TODO: Figure out how to ignore the .git directory from fetching (or at
least hashing)