kernel: sync new patch structure changes to 4.9
The 4.9 kernel works for arm64, but VCHIQ and related drivers don't
compile so they've been disabled.

Bump to 4.9.46

Move the NEWBS_DEFCONFIG selection to the inc file since it's shared
between versions.

Comment out the Makefile and scripts patches in the scc since they don't
add any code and aren't needed when building with Yocto. Leave them in
the meta layer for easier git am patching for development outside yocto.