dev-image-newbs: install kernel-devsrc
doesn't work well for external modules because some of the tools and
scripts need to be compiled for the target. Some yocto and kernel
makefile hacks could probably used to handle this.

To build an external module on the Pi, get /usr/src/kernel somewhere
read-write (or at least the scripts dir) and make scripts of modules_prepare.