userland-nographics: vc binaries without the GL stack
For aarch64 builds when vc4graphics is in MACHINE_FEATURES, mesa
provides the GL stack rather than the rpi userland recipe, which means
there's nothing to provide the vc binaries like vcgencmd.

As a workaround, make a userland-nographics version of the recipe which
builds everything and then deletes the GL stack.

This could be cleaner upstream, but I'll keep it in my layer for now.
See https://github.com/agherzan/meta-raspberrypi/issues/147