newbs-base: use squashfs zstd instead of xz
It's faster to create and much faster to decompress at runtime. UniFi
controller startup time dropped from 60 to 47 seconds, and the rest of
userspace boot may be slightly faster too.

Only downside is 8-10% larger rootfs images, but that's not too bad of
a hit and I have tons of storage space.