systemd: use RPi's serial number when possible instead of randomly generated
The end goal of this is to get persistent journald logs across reboots
with a read-only rootfs. It's easy enough to bind-mount /var/log, and
journald is smart enough to switch from /run/log when /var/log/journal
gets mounted/created.

Unfortunately this exposed a new problem, /etc/machine-id is randomized
every boot then bind-mounted over when the root filesystem is read-only.
If we bind-mount over it, the machine-id changes mid boot and the early
journal output gets lost (assigned to a "different" machine), because
machine_id_setup is run so early in boot.

Presumably /etc/machine-id could be bind-mounted by an initramfs, but
that adds a whole new layer of complexity to the build and boot process.

Fortunately, we have a unique serial number that's sent to Linux via the
device tree! All we have to do is read that, hash it up to 128 bits, and
use that as the machine-id.

Note this depends on CRYPTO_SHA256 and CRYPTO_USER_API_HASH to be 'y' in
the kernel config if booting without an initramfs.