add newbs-initramfs recipe
Using Yocto image classes requires a lot of overhead (e.g. opkg,
pre/post-process commands, etc) that I don't need for a single-file
initramfs.  Instead this recipe's Makefile compiles the source and
creates the whole initramfs.cpio.gz, so all we have to do is deploy it
as if it were an image.