nimage: don't force newbs-bootimg in IMAGE_FSTYPES
This was a problem when I tried building an initramfs as a full-blown
Yocto image (though I settled on a standalone single-binary version).

nimage still inherits newbs-bootimg to get its variables, but doesn't
mess with IMAGE_FSTYPES.

If IMAGE_FSTYPES contains nimage but not newbs-bootimg, the build will
fail at parsing because of the dependency on a non-existent task.