kernel: use schedutil governor, bring back DTC_FLAGS patch
It worked on Pi3 builds, but Pi4 builds got a broken DTB so it looks
like I do need this patch in 5.4.

Use schedutil as the default governor, demote conservative an ondemand
to modules. According to the docs, "[schedutil] generally is regarded
as a replacement for the older ondemand and conservative governors
(described below), as it is simpler and more tightly integrated with the
CPU scheduler, its overhead in terms of CPU context switches and similar
is less significant, and it uses the scheduler’s own CPU utilization
metric, so in principle its decisions should not contradict the
decisions made by the other parts of the scheduler." [1]

[1] https://www.kernel.org/doc/html/v5.4/admin-guide/pm/cpufreq.html#schedutil