gatesgarth: silence do_kernel_configcheck warnings about CONFIG_LOCALVERSION
I intentionally have CONFIG_LOCALVERSION set to -rpi in the defconfigs
(which I generate outside of Yocto generically), and
LINUX_VERSION_EXTENSION is "-newbs" for Yocto builds, so I can tell
kernel builds apart mroe easily.

In gatesgarth, this now throws warnings in do_kernel_configcheck. There
doesn't seem to be any clean way to say "don't worry it's cool for this
to be different." Thus, do an evil hack in the kern-tools script used by
configcheck.

Actually... now that I think of it, it may be easier to just make
a localversion.cfg fragment to set CONFIG_LOCALVERSION...