java: disable x11 features on icedtea/jdk/jre
libxcb-native is broken on Arch (probably related to Python 3.9) and
rather than fix it I'm just removing the need for it.

Remove x11 features from PACKAGECONFIG and dependencies for
icedtes7-native and openjdk-8-native.

Maybe if I'm lucky this speeds up build times slightly (but icedtea7 is
still slow as molasses)