add borgbackup-bin recipe
Python and all its libraries is a heavy dependency chain to add to an
image for just borgbackup, so I've been using the prebuilt binaries from
https://borg.bauerj.eu/ and putting them in ~/bin manually. With the
1.1.16 update, it wouldn't start because it couldn't dlopen
libcrypt.so.1. Fortunately we can get this with Yocto's libxcrypt-compat
recipe.

Since that required Yocto changes anyway, go ahead and make a recipe for
it that downloads the arm64 binary and installs it. I had to commit the
LICENSE file here to avoid Yocto complaining about no LIC_FILES_CHKSUM.

meta-wild-common provides a source recipe for borgbackup, so set
PREFERRED_PROVIDER to borgbackup-bin in the newbs distro conf.