conf: set preferred provider for mongodb (version 4.4)
After testing on 7.0.14 I can confirm that it does work, but it's not
worth it. Specifically, the binaries for 7.0 are twice as large as 4.4
(101M vs 48M on arm64 when stripped), and the builds seem to take longer
too. I did a benchmark on a beefy EC2 instance of a simple build and 4.4
took ~5 minutes and 7.0 took ~20. I don't have numbers for my actual
computers and these yocto builds, but I think 7.0 still takes longer.

Notes on the upgrade path, if future-me ever reads this: MongoDB has to
be updated one version at a time by bumping the Feature Compatibility
Version for each major release. This can be done by copying the db files
to a desktop and using the official prebuilt binaries. I could probably
also do it by temporarily exposing the unifi controller's db on my LAN
and connecting with mongosh from my desktop.