unifi: patch in UNPACKDIR rather than WORKDIR
I'm not sure the expected/best way to do this. UNPACKDIR is an absolute
path '${WORKDIR}/sources-unpack' and you don't want that in SRC_URI, so
this feels fragile because how are we supposed to know that
'sources-unpack' is the right name? I guess it was technically already
similarly fragile because it assumed that WORKDIR was exactly one
directory level above S.